// ft_transcendence/transcendence-backend/src/permissions/permission.enum.ts
export enum Permission {
  COMMUNITY_READ = "COMMUNITY_READ",
  COMMUNITY_UPDATE = "COMMUNITY_UPDATE",
  COMMUNITY_DELETE = "COMMUNITY_DELETE",
  MEMBER_READ = "MEMBER_READ",
  MEMBER_ADD = "MEMBER_ADD",
  MEMBER_CHANGE_ROLE = "MEMBER_CHANGE_ROLE",
  MEMBER_REMOVE = "MEMBER_REMOVE",
  CHANNEL_READ = "CHANNEL_READ",
  CHANNEL_CREATE = "CHANNEL_CREATE",
  CHANNEL_UPDATE = "CHANNEL_UPDATE",
  CHANNEL_DELETE = "CHANNEL_DELETE",
  PROJECT_READ = "PROJECT_READ",
  PROJECT_CREATE = "PROJECT_CREATE",
  PROJECT_UPDATE = "PROJECT_UPDATE",
  PROJECT_DELETE = "PROJECT_DELETE",
  TASK_READ = "TASK_READ",
  TASK_CREATE = "TASK_CREATE",
  TASK_UPDATE = "TASK_UPDATE",
  TASK_DELETE = "TASK_DELETE",
}